requirements.txt
fastapi
uvicorn
torch
numpy
pandas
faiss-cpu
fasttext
gensim
